+++ src/tasks/client-integrity-task.h
@@ -22,6 +22,7 @@
 #ifndef ClientIntegrityTask_h
 #define ClientIntegrityTask_h
 
+#include <atomic>
 #include "tasks/task.h"
 
 namespace residue {
@@ -36,8 +37,21 @@ class ClientIntegrityTask final : public Task
 {
 public:
     ClientIntegrityTask(Registry* registry, unsigned int interval);
+    void performCleanup();
+
+    inline void pauseCleanup()
+    {
+        m_performCleanUpOnSchedule = true;
+    }
+
+    inline void resumeCleanup()
+    {
+        m_performCleanUpOnSchedule = false;
+    }
 protected:
     virtual void execute() override;
+private:
+    std::atomic<bool> m_performCleanUpOnSchedule;
 };
 }
 #endif /* ClientIntegrityTask_h */