+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [Unreleased]
+### Fixes
+- Fixed issue where client integrity tasks could remove dead clients that were pending (in log queue) [issue #73]
+
 ## [1.4.1] - 12-02-2018
 ### Fixes
 - Output for `sess --stats` fixed for multiple sessions